feat: implemented author delete if not have any books

+++ src/authors/authors.service.ts
@@ -1,14 +1,24 @@
-import { Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  forwardRef,
+  Inject,
+  Injectable,
+} from '@nestjs/common';
 import { Prisma } from '@prisma/client';
 
 import { AuthorsRepository } from './authors.repository';
 import { CreateAuthorDto, UpdateAuthorDto } from './dto';
 import { Author } from './entities';
 import { getPagination } from 'src/common/utils/pagination.util';
+import { BooksService } from '../books/books.service';
 
 @Injectable()
 export class AuthorsService {
-  constructor(private repository: AuthorsRepository) {}
+  constructor(
+    private repository: AuthorsRepository,
+    @Inject(forwardRef(() => BooksService))
+    private booksService: BooksService,
+  ) {}
 
   create(data: CreateAuthorDto): Promise<Author> {
     return this.repository.create(data);
@@ -47,6 +57,13 @@ export class AuthorsService {
   }
 
   async remove(id: string): Promise<void> {
+    const hasBooks = await this.booksService.hasBooksByAuthor(id);
+    if (hasBooks) {
+      throw new BadRequestException(
+        'Author has existing books and cannot be removed.',
+      );
+    }
+
     await this.repository.remove(id);
   }
